feat(REPLRunner): add Jupynium run above/below commands
This commit introduces two new commands, JupyniumRunAbove and
JupyniumRunBelow, to the REPLRunner module. These commands allow
users to execute Jupyter notebook cells above or below the current
cell. The commands are also added to the which_key mappings for
easy access.

The commit also updates the repl_types table to include these new
commands in the keymap set for the Jupynium REPL.

BREAKING CHANGE: This modifies the behavior of the '[e' and ']e'
keys in the Jupynium REPL mode. Users who have custom bindings for
these keys may need to update their configurations.