refactor: optimize `LazyGit` function in `lvim.core.terminal`
This commit optimizes the `lazygit_toggle` function in `lvim.core.terminal`.
The function is still triggered by `<leader>gg` in `which-key`, but now it
opens the repository based on the current file's directory instead of the
default behavior of opening the working directory. If it cannot find the
repository, it will adopt the default behavior and open the working directory,
which can be used to create a new repository.

NOTE: The behavior of `<leader>gg` keybinding remains the same, but the
underlying function it triggers has been optimized.