feat(dap): add function to edit breakpoints
Added a new function `DAP_edit_breakpoint` to allow editing of existing
breakpoints. This function checks for existing breakpoints in the current
buffer and allows the user to modify the condition, hit condition, and log
message. If no matching breakpoint is found, a new one is created. Also added
a key mapping for this function in which_key and updated the mouse menu.